Update Portfolio Tracker
Added columns Transaction Value and Current Value to the portfolio DataFrame in the __init__ method.

Added automatic Transaction Value and Current Value calculations to the add_asset method.

Updated the price method: allowing the Yahoo Finance package to decide the intervals, making the method more robust for generating graphs. For large time periods, using a fixed interval of one day caused errors.

Added the graph method to present graphs of the portfolio or a selection of tickers. When no tickers are given to the method, the paths of all stocks in the portfolio are shown. Otherwise, the paths of the given tickers are displayed.

Added the __str__ method to make it easier for the user to print the portfolio.

Removed the maximum on the number of printed columns to ensure the user can see all required columns related to the portfolio.