Created the total portolio value method and implemented small adjustments
Made a small adjustment in the yfinance downloads: added progress = False to hide progress messages..

Made a small adjustment to the __str__ method: added \n before and after the dataframe and rounded the values, to increase the readability of the printed messages. 

Added a method portfolio_value. This can be used to see the total portfolio value as well as the share of the total portfolio value per asset, per asset class, and per sector.

Work that still needs to be done:
Adding a method to make a five year forecast of the portfolio, add comments and user explanations to the methods and write a text file that explains how to use the file.