Added two methods needed for portfolio forecasting
Added method S_T, which can be used to generate a path of a stock. The method assumes the stock to move according to a Geometric Brownian Motion. 

Added method portfolio_forecast. This method creates a forecast of the portfolio. In order to do so, it makes use of the method S_T. The method generates stock paths according using the method S_T, with the drift term and volitility term corresponding to the historic values of that specific stock. The method is not yet finished. It now creates the paths of all the stocks in the portfolio, but we still need to make a weighted combination of the stocks into one path.