Finished portfolio_forecast method & made small adjustments to graph method
Finished portfolio_forecast method: Made sure that the portfolio_forecast method combines the forecasts of different stocks in the portfolio by creating a weighted sum, based on the quantities in the portfolio. Added a graph section in the method, which allows users to get a visual representation of the forecast.

Made small adjustments to the graph method: added tight_layout for small improvements in the layout. Added xlabel, ylabel and a title to the plot.